conf: Separate component
